Split trap frame into it's own header
