+++ arch/aarch64/thread.hpp
@@ -1,4 +1,5 @@
 #include <kern_datatypes.hpp>
+#include <arch/aarch64/trap.hpp>
 
 #ifndef _PANTHEON_ARM_THREAD_HPP_
 #define _PANTHEON_ARM_THREAD_HPP_
@@ -119,51 +120,6 @@ typedef struct CpuContext
 
 }CpuContext;
 
-typedef enum PSTATEMode : UINT64
-{
-	PSTATE_MODE_USER = 0x00,
-	PSTATE_MODE_KERN = 0x04,
-	PSTATE_MODE_HYPER = 0x08,
-	PSTATE_MODE_FIRM = 0x0C,
-}PSTATEMode;
-
-typedef struct TrapFrame
-{
-	UINT64 Regs[31];
-	UINT64 PC;
-	PSTATEMode PSTATE;
-	UINT64 ESR;
-	UINT64 SP;
-	UINT64 PADDING;
-
-	VOID Wipe()
-	{
-		for (UINT64 &Reg : Regs)
-		{
-			Reg = 0;
-		}
-		PC = 0;
-		PSTATE = PSTATE_MODE_USER;
-		SP = 0;
-	}
-
-	VOID SetUser()
-	{
-		this->PSTATE = PSTATE_MODE_USER;
-	}
-
-	VOID SetKern()
-	{
-		this->PSTATE = PSTATE_MODE_KERN;
-	}
-
-	UINT64 &GetIntArgument(UINT8 Index)
-	{
-		return this->Regs[Index];
-	}
-}TrapFrame;
-
-static_assert(sizeof(TrapFrame) == (8ULL *36ULL));
 
 }
